+++ ProjetoOperacoes/ProjetoOperacoes/Models/AccomplishedApplication.cs
@@ -2,7 +2,7 @@
 {
     public class AccomplishedApplication : ApplicationModels
     {
-        public AccomplishedApplication(string id, string idTiposContas, string descricao, bool parcelas, decimal valorIndividual, decimal valorTotal) : base(id, idTiposContas, descricao, parcelas, valorIndividual, valorTotal)
+        public AccomplishedApplication(string id, string idTiposContas, string descricao, bool parcelas, double valorIndividual, double valorTotal) : base(id, idTiposContas, descricao, parcelas, valorIndividual, valorTotal)
         {
 
         }
