+++ ProjetoOperacoes/ProjetoOperacoes/Models/FutureApplicationModel.cs
@@ -2,7 +2,7 @@
 {
     public class FutureApplicationModel : ApplicationModels
     {
-        public FutureApplicationModel(string id, string idTiposContas, string descricao, bool parcelas, decimal valorIndividual, decimal valorTotal) : base(id, idTiposContas, descricao, parcelas, valorIndividual, valorTotal)
+        public FutureApplicationModel(string id, string idTiposContas, string descricao, bool parcelas, double valorIndividual, double valorTotal) : base(id, idTiposContas, descricao, parcelas, valorIndividual, valorTotal)
         {
 
         }
