+++ ProjetoOperacoes/ProjetoOperacoes/Models/ProgressApplicationModel.cs
@@ -6,7 +6,7 @@
         public int PayInstallments { get; set; }
         public int NumberInstallments { get; set; }
 
-        public ProgressApplicationModel(string id, string idTiposContas, string descricao, bool parcelas, decimal valorIndividual, decimal valorTotal) : base(id, idTiposContas, descricao, parcelas, valorIndividual, valorTotal)
+        public ProgressApplicationModel(string id, string idTiposContas, string descricao, bool parcelas, double valorIndividual, double valorTotal) : base(id, idTiposContas, descricao, parcelas, valorIndividual, valorTotal)
         {
 
         }
